Fix revision offsets in graph.
I noticed that test_pcint.bin had been dropped from the Makefile
(merge error). After adding it again, it was shown with incorrect values on
the x-axis (it was newly in recent_sizes.json,but the offset calculation was
too simplistic)

Fixed the bug and rewrote sizes.template a bit.

Also added the make target 'history'. Couldn't resist the pun.